refactored new file utils
