+++ app/react/Library/actions/saveEntityWithFiles.ts
@@ -2,6 +2,7 @@ import superagent from 'superagent';
 import { Dispatch } from 'redux';
 import { ClientEntitySchema, ClientFile } from 'app/istore';
 import { ensure } from 'shared/tsUtils';
+import { constructFile } from 'shared/fileUploadUtils';
 import * as attachmentsTypes from 'app/Attachments/actions/actionTypes';
 
 export const readFileAsBase64 = async (file: Blob, cb: (file: any) => void) =>
@@ -15,15 +16,6 @@ export const readFileAsBase64 = async (file: Blob, cb: (file: any) => void) =>
     };
     reader.readAsDataURL(file);
   });
-
-export const constructFile = ({ serializedFile: base64, originalname }: ClientFile) => {
-  const fileParts = base64!.split(',');
-  const fileFormat = fileParts[0].split(';')[0].split(':')[1];
-  const fileContent = fileParts[1];
-  const buff = Buffer.from(fileContent, 'base64');
-
-  return new File([buff], originalname || '', { type: fileFormat });
-};
 
 export const saveEntityWithFiles = async (entity: ClientEntitySchema, dispatch?: Dispatch<{}>) =>
   new Promise((resolve, reject) => {
